+++ src/main/java/org/pickaid/picontent/api/clientdecl/PiContentRig.java
@@ -1,0 +1,10 @@
+package org.pickaid.picontent.api.clientdecl;
+
+public final class PiContentRig {
+    private PiContentRig() {
+    }
+
+    public static PiClientDeclaration geoItem(String rigId) {
+        return new PiClientDeclaration("rig", rigId);
+    }
+}