+++ src/main/java/org/pickaid/picontent/api/item/PiItemProperties.java
@@ -1,0 +1,10 @@
+package org.pickaid.picontent.api.item;
+
+public final class PiItemProperties {
+    private PiItemProperties() {
+    }
+
+    public static PiItemPropertyDeclaration using() {
+        return new PiItemPropertyDeclaration("", "using");
+    }
+}